remove old ocr
