+++ autowsgr/game/relative_location.yaml
