+++ autowsgr/timer/controllers/android_controller.py
@@ -137,11 +137,12 @@ class AndroidController:
             raise ValueError("arg 'delay' should be positive or 0")
         x1, y1 = relative_to_absolute((x1, y1), self.resolution)
         x2, y2 = relative_to_absolute((x2, y2), self.resolution)
-        input_str = f"input swipe {str(x1)} {str(y1)} {str(x2)} {str(y2)} {int(duration * 1000)}"
+        duration = int(duration * 1000)
+        input_str = f"input swipe {str(x1)} {str(y1)} {str(x2)} {str(y2)} {duration}"
         if self.config.SHOW_ANDROID_INPUT:
             self.logger.debug(input_str)
         self.shell(input_str)
-        time.sleep(duration + delay)
+        time.sleep(delay)
 
     def swipe(self, x1, y1, x2, y2, duration=0.5, delay=0.5, *args, **kwargs):
         """匀速滑动模拟器相对坐标 (x1,y1) 到 (x2,y2).
